# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  baby-backend_internal_models.EntryComment:
    properties:
      author_avatar:
        type: string
      author_name:
        type: string
      entry_id:
        type: integer
      id:
        type: integer
      text:
        type: string
      timestamp:
        type: string
      user_id:
        type: integer
    type: object
...
  baby-backend_internal_models.TimelineEntry:
    properties:
      author_avatar:
        type: string
      author_id:
        type: integer
      author_name:
        type: string
      baby_id:
        type: integer
      comments:
        description: Relations (loaded via Preload)
        items:
          $ref: '#/definitions/baby-backend_internal_models.EntryComment'
        type: array
      created_at:
        type: string
      date:
        type: string
      description:
        type: string
      featured:
        type: boolean
      id:
        type: integer
      image_url:
        type: string
      images:
        items:
          type: string
        type: array
      likes:
        type: integer
      milestone_title:
        type: string
      tags:
        items:
          type: string
        type: array
      type:
        description: '"photo" | "video" | "text" | "milestone"'
        type: string
      updated_at:
        type: string
    type: object
...
  baby-backend_internal_services.CreateEntryInput:
    properties:
      baby_id:
        type: integer
      date:
        type: string
      description:
        type: string
      image_url:
        type: string
      images:
        items:
          type: string
        type: array
      milestone_title:
        type: string
      tags:
        items:
          type: string
        type: array
      type:
        enum:
        - photo
        - video
        - text
        - milestone
        type: string
    required:
    - baby_id
    - date
    - type
    type: object
...
  baby-backend_internal_utils.APIResponse:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: 宝宝时光 · 家庭馆 — 宝宝成长记录与家庭共享后端 API
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://github.com
  title: Baby Backend API
  version: "1.0"
paths:
...
  /timeline:
    get:
      description: 按 baby_id 查询时间线动态，支持筛选：type=all|featured|member|content，member
        时需传 name，content 时需传 content_type。
      parameters:
      - description: 宝宝 ID
        in: query
        name: baby_id
        required: true
        type: integer
      - description: '筛选类型: all | featured | member | content'
        in: query
        name: type
        type: string
      - description: 当 type=member 时，按成员名筛选
        in: query
        name: name
        type: string
      - description: '当 type=content 时，内容类型: photo | video | text | milestone'
        in: query
        name: content_type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/baby-backend_internal_utils.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/baby-backend_internal_models.TimelineEntry'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: 时间线列表
      tags:
      - Timeline
    post:
      consumes:
      - application/json
      parameters:
      - description: 动态内容
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/baby-backend_internal_services.CreateEntryInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/baby-backend_internal_utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/baby-backend_internal_models.TimelineEntry'
              type: object
      security:
      - BearerAuth: []
      summary: 发布动态
      tags:
      - Timeline
  /timeline/{id}:
    delete:
      parameters:
      - description: 动态 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/baby-backend_internal_utils.APIResponse'
      security:
      - BearerAuth: []
      summary: 删除动态
      tags:
      - Timeline
    get:
      parameters:
      - description: 动态 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/baby-backend_internal_utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/baby-backend_internal_models.TimelineEntry'
              type: object
      security:
      - BearerAuth: []
      summary: 单条动态详情
      tags:
      - Timeline
  /timeline/{id}/comments:
    post:
      consumes:
      - application/json
      parameters:
      - description: 动态 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 评论内容
        in: body
        name: input
        required: true
        schema:
          properties:
            text:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/baby-backend_internal_utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/baby-backend_internal_models.EntryComment'
              type: object
      security:
      - BearerAuth: []
      summary: 添加评论
      tags:
      - Timeline
  /timeline/{id}/comments/{commentId}:
    delete:
      parameters:
      - description: 动态 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 评论 ID
        in: path
        name: commentId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/baby-backend_internal_utils.APIResponse'
      security:
      - BearerAuth: []
      summary: 删除评论
      tags:
      - Timeline
  /timeline/{id}/featured:
    patch:
      parameters:
      - description: 动态 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/baby-backend_internal_utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/baby-backend_internal_models.TimelineEntry'
              type: object
      security:
      - BearerAuth: []
      summary: 切换精选状态
      tags:
      - Timeline
  /timeline/{id}/like:
    post:
      description: 同一条动态重复调用会在点赞 / 取消之间切换。返回当前 liked 状态。
      parameters:
      - description: 动态 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/baby-backend_internal_utils.APIResponse'
      security:
      - BearerAuth: []
      summary: 点赞 / 取消点赞
      tags:
      - Timeline
securityDefinitions:
  BearerAuth:
    description: 'JWT Bearer Token. Format: "Bearer <token>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"